backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9

# Validation & settings
pydantic[email]==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
python-multipart==0.0.9

# Auth — local
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-jose[cryptography]==3.3.0
itsdangerous==2.2.0

# Auth — Microsoft Entra ID (OAuth2/OIDC)
authlib==1.3.0
httpx==0.27.0

# Security
slowapi==0.1.9          # rate limiting

# Logging & monitoring
structlog==24.1.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
